Move ui package to the views
Make sense it being there
