+++ smssync/src/main/AndroidManifest.xml
@@ -40,11 +40,11 @@
             android:theme="@style/AppTheme">
         <!-- messages tab -->
         <activity
-                android:name=".presentation.ui.activity.MainActivity"
+                android:name=".presentation.view.ui.activity.MainActivity"
                 android:label="@string/app_name"
                 android:theme="@style/AppTheme.NoActionBar"
                 android:launchMode="singleTask"
-                android:parentActivityName=".presentation.ui.activity.MainActivity">
+                android:parentActivityName=".presentation.view.ui.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
@@ -52,10 +52,10 @@
             </intent-filter>
             <meta-data
                     android:name="android.support.PARENT_ACTIVITY"
-                    android:value=".presentation.ui.activity.MainActivity"/>
+                    android:value=".presentation.view.ui.activity.MainActivity"/>
         </activity>
         <activity
-                android:name=".presentation.ui.activity.OauthActivity"
+                android:name=".presentation.view.ui.activity.OauthActivity"
                 android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW"/>
@@ -64,13 +64,13 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".presentation.ui.activity.SettingsActivity"
+                android:name=".presentation.view.ui.activity.SettingsActivity"
                 android:label="@string/action_settings"
                 android:exported="false"
                 android:launchMode="singleTask"
-                android:parentActivityName=".presentation.ui.activity.MainActivity"/>
+                android:parentActivityName=".presentation.view.ui.activity.MainActivity"/>
         <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".presentation.ui.activity.MainActivity"/>
+                android:value=".presentation.view.ui.activity.MainActivity"/>
     </application>
 </manifest>
