+++ smssync/src/main/java/org/addhen/smssync/presentation/view/ui/adapter/LogAdapter.java
@@ -10,49 +10,54 @@
  * packaging of this file. Please review the following information to
  * ensure the GNU Lesser General Public License version 3 requirements
  * will be met: http://www.gnu.org/licenses/lgpl.html.
- *
+ *  
  * If you have questions regarding the use of this file, please contact
  * Ushahidi developers at team@ushahidi.com.
  */
 
-package org.addhen.smssync.presentation.ui.adapter;
+package org.addhen.smssync.presentation.view.ui.adapter;
 
 import com.addhen.android.raiburari.presentation.ui.adapter.BaseRecyclerViewAdapter;
 
 import org.addhen.smssync.R;
-import org.addhen.smssync.presentation.model.FilterModel;
+import org.addhen.smssync.presentation.model.LogModel;
 
+import android.support.v7.widget.AppCompatTextView;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import java.util.List;
 
-import butterknife.Bind;
 import butterknife.ButterKnife;
+import butterknife.Bind;
 
 /**
  * @author Ushahidi Team <team@ushahidi.com>
  */
-public class FilterAdapter extends BaseRecyclerViewAdapter<FilterModel> {
+public class LogAdapter extends BaseRecyclerViewAdapter<LogModel> {
 
     private View mEmptyView;
 
-    public FilterAdapter(final View emptyView) {
+    public LogAdapter(final View emptyView) {
         mEmptyView = emptyView;
         onDataSetChanged();
     }
 
     @Override
     public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup viewGroup) {
-        return new Widgets(LayoutInflater.from(viewGroup.getContext())
-                .inflate(R.layout.filter_list_item, viewGroup, false));
+        return new Widgets(LayoutInflater.from(viewGroup.getContext()).inflate(
+                R.layout.list_message_item, viewGroup, false));
     }
 
     @Override
-    public void setItems(List<FilterModel> items) {
+    public int getAdapterItemCount() {
+        return getItems().size();
+    }
+
+    @Override
+    public void setItems(List<LogModel> items) {
         super.setItems(items);
         onDataSetChanged();
     }
@@ -66,24 +71,20 @@ public class FilterAdapter extends BaseRecyclerViewAdapter<FilterModel> {
     }
 
     @Override
-    public int getAdapterItemCount() {
-        return getItems().size();
-    }
-
-    @Override
     public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
-        ((Widgets) holder).phoneNumber.setText(getItem(position).phoneNumber);
+        final LogModel logModel = getItem(position);
+        // initialize view with content
+        ((Widgets) holder).logMessage.setText(logModel.message);
     }
 
     public class Widgets extends RecyclerView.ViewHolder {
 
-        @Bind(R.id.filter_phone_number)
-        TextView phoneNumber;
+        @Bind(R.id.log_message)
+        AppCompatTextView logMessage;
 
         public Widgets(final View view) {
             super(view);
             ButterKnife.bind(this, view);
-
         }
     }
 }
