+++ smssync/src/main/java/org/addhen/smssync/presentation/view/ui/fragment/IntegrationFragment.java
@@ -15,13 +15,13 @@
  * Ushahidi developers at team@ushahidi.com.
  */
 
-package org.addhen.smssync.presentation.ui.fragment;
+package org.addhen.smssync.presentation.view.ui.fragment;
 
 import com.addhen.android.raiburari.presentation.ui.fragment.BaseFragment;
 
 import org.addhen.smssync.R;
 import org.addhen.smssync.presentation.App;
-import org.addhen.smssync.presentation.ui.activity.MainActivity;
+import org.addhen.smssync.presentation.view.ui.activity.MainActivity;
 
 import butterknife.OnClick;
